+++ package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "build": "tsc",
     "test": "jest",
-    "dev": "nodemon src/server.ts"
+    "dev": "nodemon src/server/server.ts"
   },
   "devDependencies": {
     "@types/hapi__joi": "^17.1.4",
