+++ src/game/rules.ts
