+++ src/models/card.ts
